requirement.txt
Flask
opencv-python-headless
numpy
torch
torchvision
Pillow